# docs/swagger.yaml
basePath: /
definitions:
  main.BridgeResponse:
    properties:
      bridge:
        type: string
      interface:
        type: string
    type: object
  main.InterfaceRequest:
    properties:
      bridge1:
        type: string
      bridge2:
        type: string
    type: object
  main.SliceRequest:
    properties:
      DstIP:
        type: string
      FlowRate:
        type: integer
      SliceSD:
        type: string
      SrcIP:
        type: string
    type: object
  main.VxlanInterfaceRequest:
    properties:
      bindInterface:
        type: string
      localBrIp:
        type: string
      remoteIp:
        description: "LocalBridgeName\t\tstring\t`json:\"localBrName\"`"
        type: string
      vxlanId:
        type: string
      vxlanInterface:
        type: string
    type: object
info:
  contact: {}
  description: API endpoints for managing bridges and interfaces.
  title: Bridge API
  version: "1.0"
paths:
  /api/v1/bridge:
    get:
      description: Get the current bridge and its connected interface
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.BridgeResponse'
      summary: Get current bridge and connected interface
      tags:
      - bridge
  /api/v1/bridge/{bridge_name}:
    post:
      consumes:
      - application/json
      description: Add a new bridge with the given name
      parameters:
      - description: Bridge name
        in: path
        name: bridge_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Bridge created successfully
          schema:
            type: string
        "400":
          description: Invalid bridge name
          schema:
            type: string
      summary: Add a new bridge
      tags:
      - bridge
  /api/v1/interface:
    post:
      consumes:
      - application/json
      description: Add a new interface between two bridges
      parameters:
      - description: Interface request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.InterfaceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Interface added successfully
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            type: string
      summary: Add a new interface
      tags:
      - interface
  /api/v1/slice/{bridge_name}:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bridge name
        in: path
        name: bridge_name
        required: true
        type: string
      - description: Slice request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.SliceRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Slice Installed
          schema:
            type: string
      summary: Add slice on interface
      tags:
      - slice
  /api/v1/slice/{bridge_name}/{slice_sd}}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Bridge name
        in: path
        name: bridge_name
        required: true
        type: string
      - description: Slice SD identifier
        in: path
        name: slice_sd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Slice deletion successful
          schema:
            type: string
      summary: Del slice on interface
      tags:
      - slice
  /api/v1/vxlan/{bridge_name}:
    delete:
      consumes:
      - application/json
      description: Delete a exist bridge with the given name
      parameters:
      - description: Bridge name
        in: path
        name: bridge_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Bridge delete successfully
          schema:
            type: string
      summary: Delete bridge
      tags:
      - vxlan
    post:
      consumes:
      - application/json
      description: Add a new bridge with the given name
      parameters:
      - description: Bridge name
        in: path
        name: bridge_name
        required: true
        type: string
      - description: Vxlan Interface request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.VxlanInterfaceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Bridge created successfully
          schema:
            type: string
        "400":
          description: Invalid bridge name
          schema:
            type: string
      summary: Add a new bridge
      tags:
      - vxlan
  /api/v1/vxlan/{bridge_name}/activate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bridge name
        in: path
        name: bridge_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Bridge Activated
          schema:
            type: string
        "400":
          description: Invalid bridge name
          schema:
            type: string
      summary: '[Deprecated] Activate vxlan bridge'
      tags:
      - vxlan
swagger: "2.0"